% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfaandb}
\alias{dfaandb}
\title{DATASET OF ALBRETCH AND BAILEY}
\format{
A data frame with 42 rows and three variables:
\describe{
\item{effort}{Effort measured in -------.}
\item{size}{Size measured in ....}
\item{source}{A or B indicating one source or another.}
}
}
\source{
\url{https://....domain.com... }
}
\usage{
dfaandb
}
\description{
Data of effort and size for several projects
}
\details{
No missing values
A dataset containing -----  whatever you put here.
}
\keyword{datasets}
